append "Safe" to risky upstream method names
Some upstream methods are used to check if unlocking is possible when it
is known that a biometric auth has succeeded, but they don't and can't
take into account biometric second factor. Renaming these methods gives
us build errors for future callers, so we can make sure they are safe.