[temporary] SettingsLib: add battery health to BatteryStatus
When 16QPR1 is pushed, if there is an upstream commit that adds battery health, drop this commit in
favor of the upstream commit.

This addition first appeared in the July 2025's SystemUIGoogle in the stock OS images. As
BatteryStatus is in AOSP, this may conflict when upstream pushes tags, hence the [temporary] tag.
However, note that battery health is nothing new in terms of BatteryManager constants.