avoid NPE during package uninstall
On Android 16, when multiple package installer sessions try to install the same
apk, the resulting ap install is temporarily broken until a reboot is done. If
app uninstall was attempted, while the package is in this broken state, the
package state is null for that package, which leads to system crash.

Avoid the NPE for now until a fix is implemented to avoid ephemeral broken installs.