force default unsupported notification adjustments
The `getUnsupportedAdjustments` is meant to be a read-only method to get unsupported adjustment
types. However, the keys in `mDefaultUnsupportedAdjustments` are effectively only used when
`setNasUnsupportedDefaults` is called.

`setNasUnsupportedDefaults` can only be called when a notification assistant is granted or revoked
access. (Can also be called on a backup restore, but that's unlikely). The resulting behavior is
that the default unsupported list is only applied when the user toggles Settings > Notifications >
Enhanced notifications.

Force the default unsupported notification adjustments to be read if the list is empty.